+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### February
 
+* February 28 - Don't show welcome notifications in UI #790 @gathuku
 * February 27 - Fix Preview image doesn't change when select new avatar in profile page #787 @ryy
 * February 26 - Business welcome emails #772 @accua
 * February 18 - Add "referred by" to admin emails #780 @PatrickMcSweeny